requirements.txt
streamlit
google-generativeai
python-dotenv
requests
pandas
pillow
streamlit-authenticator
python-dotenv
PyPDF2